On tab completion, UI thread is blocked
Root cause: Synchronous call to Runspace.Invoke is made from the UI thread.
Unlike other cmdlets, TabCompletion returns a collection of strings and it
 should be used in the UI thread

On tab completion, ActivePackageSource is not updated. So, performing tab completion after a change in ActivePackageSource will still use the old one
a.	Root cause: PrivateData on the Host is set with the ActivePackageSource before invoking cmdlets. Tab completion, however, needs to return the value to the WpfConsole layer
b.	Solution: Set the Private Data on Tab completion as well